requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.5
google-auth==2.23.4
google-oauth2-tool==0.0.3
Pillow==10.1.0
python-dateutil==2.8.2
